Update README.md
PROBLEM STATEMENT:


To perform classification analysis on Iris dataset. Perform any two classification algorithms and compare the accuracy.

Introduction
Classification is one of the most basic yet fundamental applications of machine learning. It is a sub-category of SUPERVISED LEARNING (Supervised learning is the process wherein the value or result that we want to predict is within the training data (independent variable) and the value we which is on the data that we want to predict is called The dependent variable.)  where you basically categories your output into two classes. An example can be a model that predicts the outcome of a training set as either yes or no, good or bad, accepted or declined or may be true or false.
While this to a large extent explains the fundamental theory behind classification, there are cases wherein the independent variable contains more than two classes, where you can’t necessarily divide the output into yes or no’s.
This is where we make use of a phenomenon called  Multi-Class Classification. In multi-class classification, you can basically divide the instances of your output data set into more than two classes. 

Brief 
We have been asked to perform a classification analysis on the Iris dataset. In the dataset, we have been given Sepal values and Petal values.

The four features that were given are:

1.Sepal length
2.Sepal width
3.Petal length 
4.Petal width
Now, Depending on these features there are basically three (3) types of flowers
They are:
1.Iris - Versicolor
2.Iris - Setosa
3.Iris Virginica

What are to do is to train the dataset so that with the given features, our model can predict the class our flower belongs to. 



TOOLS
The tools used in this project are:
1.Python: Programming language used to perform the analysis 
2.Google colab Notebook

Algorithms 

1.KNN Algorithm: K Nearest Neighbor belongs to the supervised learning domain and it is a simple algorithm that basically assumes that similar things would be in close distance /proximity of each other. The main idea is that if a point in a data set is close to another point then it is safe to assume they belong to similar classes.


Decision Tree: This is also a very powerful tool used for classification and also predictions. It is essentially a tree-like model that shows different decisions and what the consequences of those decisions would be. The internal nodes denote a test case on an attribute and each branch shows the outcome or result of the test, and each leaf node holds a class.

Conclusion
In the course of this project, I have successfully implemented the KNN algorithm and Decision Tree classifiers for the IRIS dataset. I normalised my data set while preparing the data in order to bring the values of the data to a common scale thereby increasing the accuracy of the data. I also used the fit method to train my model. I got an accuracy of 100% on both the KNN algorithm and Decision Tree, I also computed the confusion matrix of the prediction.


Finally, I visualised the data using Matplotlib and Seaborn and was able to derive insights from the data


Kindly access the link below for my google colab notebook

https://colab.research.google.com/drive/1ZTzITFjhNQugs-nu3ONJIMgcZmMUC8Bn?usp=sharing